game.rs and project structure

+++ src/settings.rs
@@ -1,3 +1,3 @@
-pub const FIELD_WIDTH: u8 = 10;
-pub const FIELD_HEIGHT: u8 = 10;
-pub const MINE_COUNT: u8 = 10;
+pub const FIELD_WIDTH: usize = 10;
+pub const FIELD_HEIGHT: usize = 10;
+pub const MINE_COUNT: usize = 10;
